refactor: move view binding to viewmodel
