+++ vs-commitizen.vs2015/VsCommitizenView.xaml.cs
@@ -1,216 +1,23 @@
-﻿using Microsoft.TeamFoundation.MVVM;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Runtime.CompilerServices;
+﻿using System;
 using System.Windows.Controls;
-using System.Windows.Input;
-using vs_commitizen.vs.Models;
-using vs_commitizen.vs.Extensions;
-using System.Text.RegularExpressions;
-using vs_commitizen.vs.Interfaces;
+using vs_commitizen.vs.ViewModels;
 
 namespace vs_commitizen.vs
 {
     /// <summary>
     /// Interaction logic for VsCommitizenView.xaml
     /// </summary>
-    public partial class VsCommitizenView : UserControl, INotifyPropertyChanged, ICommentBuilder
+    public partial class VsCommitizenView : UserControl
     {
-        #region Bound properties
-        private List<CommitType> _commitTypes;
-        public List<CommitType> CommitTypes
-        {
-            get => _commitTypes;
-            set
-            {
-                _commitTypes = value;
-                OnPropertyChanged();
-            }
-        }
-
-        private CommitType _selectedCommitType;
-        public CommitType SelectedCommitType
-        {
-            get => _selectedCommitType;
-            set
-            {
-                _selectedCommitType = value;
-                OnPropertyChanged();
-                OnPropertyChanged(nameof(OnProceed));
-                OnPropertyChanged(nameof(SubjectLength));
-            }
-        }
-
-        private string _scope;
-        public string Scope
-        {
-            get => _scope;
-            set
-            {
-                _scope = value;
-                OnPropertyChanged();
-                OnPropertyChanged(nameof(SubjectLength));
-
-            }
-        }
-
-        private string _body;
-        public string Body
-        {
-            get => _body;
-            set
-            {
-                _body = value;
-                OnPropertyChanged();
-            }
-        }
-
-        private string _breakingChanges;
-        public string BreakingChanges
-        {
-            get => _breakingChanges;
-            set
-            {
-                _breakingChanges = value;
-                OnPropertyChanged();
-            }
-        }
-
-        private string _issuesAffected;
-        public string IssuesAffected
-        {
-            get => _issuesAffected;
-            set
-            {
-                _issuesAffected = value;
-                OnPropertyChanged();
-            }
-        }
-
-        private string _subject;
-        public string Subject
-        {
-            get => _subject;
-            set
-            {
-                _subject = value;
-                OnPropertyChanged();
-                OnPropertyChanged(nameof(OnProceed));
-                OnPropertyChanged(nameof(SubjectLength));
-            }
-        }
-
-        public int SubjectLength
-        {
-            get
-            {
-                int type =this.SelectedCommitType?.Type.Length ?? 0;
-                int scope = this.Scope?.Length ?? 0;
-                int subject = this.Subject?.Length ?? 0;
-                var sum = type + scope + subject;
-                sum = sum == 0 ? 0 : sum + 1;
-
-                return sum;
-            }
-        }
-
-        private bool _hasGitPendingChanges;
-        public bool HasGitPendingChanges
-        {
-            get => _hasGitPendingChanges;
-            set
-            {
-                _hasGitPendingChanges = value;
-                OnPropertyChanged();
-                OnPropertyChanged(nameof(OnProceed));
-            }
-        }
-        #endregion
+        CommitizenViewModel _viewModel = new CommitizenViewModel();
 
         public VsCommitizenView()
         {
             InitializeComponent();
 
-            this.CommitTypes = new List<CommitType>
-            {
-                new CommitType("feat", "A new feature"),
-                new CommitType("fix", "A bug fix"),
-                new CommitType("docs", "Documentation only changes"),
-                new CommitType("style", "Changes that do not affect the meaning of the code (formatting, etc)"),
-                new CommitType("refactor", "A code change that neither fixes a bug nor adds a feature"),
-                new CommitType("perf", "A code change that improves performance"),
-                new CommitType("test", "Adding missing tests or correcting existing tests"),
-                new CommitType("build", "Changes that affect the build system or external dependencies (example scopes: gulp, etc)"),
-                new CommitType("ci", "Changes to our CI configuration files and scripts (example scopes: Travis, etc)"),
-                new CommitType("chore", "Other changes that don't modify src or test files"),
-                new CommitType("revert", "Reverts a previous commit")
-            };
-            this.OnProceed = new RelayCommand(Proceed, CanProceed);
-            this.HasGitPendingChanges = true;   //TODO: Correct way to bind this
-
-            this.DataContext = this;
+            this.DataContext = _viewModel;
         }
 
-        private void OnPropertyChanged([CallerMemberName] string propertyName = null)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        }
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        public bool CanProceed(object param)
-        {
-            if (this.cbType.SelectedIndex < 0) return false;
-            if (string.IsNullOrWhiteSpace(this.tbxSubject.Text)) return false;
-            if (!this.HasGitPendingChanges) return false;
-
-            return true;
-        }
-
-        public void Proceed(object param)
-        {
-            var doCommit = param is bool && (bool)param;
-            ProceedExecuted?.Invoke(this, doCommit);
-        }
-
-        public string GetComment()
-        {
-            var hasScope = !string.IsNullOrWhiteSpace(this.Scope);
-            var scope = hasScope ? $"({this.Scope.SafeTrim()})" : string.Empty;
-            var commitType = (this.cbType.SelectedItem as CommitType).Type;
-
-            var head = $"{commitType}{scope}: {this.Subject.SafeTrim()}"; //TODO: add a max size check
-            var body = string.Join("\n", this.Body.SafeTrim().ChunkBySize(100));
-
-            var hasBreakingChanges = !string.IsNullOrEmpty(this.BreakingChanges);
-            var breakingChanges = this.BreakingChanges.SafeTrim();
-            if (hasBreakingChanges)
-            {
-                breakingChanges = "BREAKING CHANGES: " + Regex.Replace(this.BreakingChanges, "^BREAKING CHANGES: ", string.Empty, RegexOptions.IgnoreCase);
-                breakingChanges = string.Join("\n", breakingChanges.ChunkBySize(100));
-            }
-
-            //TODO: prefix by #
-            var issues = !string.IsNullOrWhiteSpace(this.IssuesAffected) ? string.Join("\n", this.IssuesAffected.SafeTrim().ChunkBySize(100)) : string.Empty;
-
-            var comment = string.Empty;
-            comment += $"{head}\n\n{body}\n\n";
-            if (!string.IsNullOrEmpty(breakingChanges)) comment += $"{breakingChanges}\n\n";
-            if (!string.IsNullOrEmpty(issues)) comment += $"{issues}\n\n";
-            return comment;
-        }
-
-        public event EventHandler<bool> ProceedExecuted;
-
-        private ICommand _onProceed;
-        public ICommand OnProceed
-        {
-            get => _onProceed;
-            set
-            {
-                _onProceed = value;
-                OnPropertyChanged();
-            }
-        }
+        public CommitizenViewModel ViewModel => _viewModel;
     }
 }
